Properly initiate the data section macro.
Since the data section macro was not being properly initialized from the p env's riscv.h file, there was no begin_signature symbol in the final binary file when compiled in the virtual memory mode. The fevsr looks for this symbol to identify the memory segment which needs to be dumped. Due to this reason, there was no actual signature dump compare in the torture tests. It was erroneously comparing the dumps of the last correctly compiled prog's dump or simply comparing an empty string if no prev dump was found.